Port to AMD MI250X (#63)
* Elimintate USE_OPENMP_NO_GPU build flag.  We can't think of a
reasonable use case, and it just complicates the code.
Now HAVE_OPENMP_TARGET will imply running on GPUs.

* Enable the openMP offload build to work on AMD.

* We need to add this declare target around the MC_Base_Particle class
to compile successfully with OpenMP with either AMD clang or CCE.
However, XLC builds the code with no problem and I think the declare
target should be implicit.  I will be filing a ticket on this.

* Finally finished getting everything working with the hip port.  I
totally re-wrote the atomics code for portability between cuda and hip
builds.  It now uses inline functions instead of macros.  I've
verified that the code builds on CTS-1, CTS-2, ATS-2, and EAS-3.  The
single node smoke test gets the same results on all.

+++ src/MC_Cell_State.hh
@@ -20,7 +20,7 @@ class MC_Cell_State
    double  _cellNumberDensity;         // number density of ions in cel
 
    uint64_t _id;
-   unsigned _sourceTally;
+   uint64_t _sourceTally;
    
    MC_Cell_State();
